Converting provinces, another part. (#39)
* county locs.

* converting rulers.

* Previous holders, monarch names and regnalness.

* Province import.

* dead fixes.

* clang.

* Mathematical ignoramus.

* not really working.

* Broken savegame dejure liege recovery.

* fixing countyweight calc to be dynamic.

* fixing codacy.

* world processing.

* codacy love.

* Flop the FAP.

* some opportune moves.

* post-processing.

* le output.

* travis as well.

* fixing provinces.

* Fixing provinces and some of my overthinking.

* some pr fixes

* Other pr fixes.

+++ CK3toEU4/Source/CK3World/CoatsOfArms/CoatsOfArms.cpp
@@ -15,7 +15,7 @@ CK3::CoatsOfArms::CoatsOfArms(std::istream& theStream)
 void CK3::CoatsOfArms::registerKeys()
 {
 	registerRegex(R"(\d+)", [this](const std::string& coaID, std::istream& theStream) {
-		auto newCoA = std::make_shared<CoatOfArms>(theStream, std::stoi(coaID));
+		auto newCoA = std::make_shared<CoatOfArms>(theStream, std::stoll(coaID));
 		coats.insert(std::pair(newCoA->getID(), newCoA));
 	});
 	registerKeyword("coat_of_arms_manager_database", [this](const std::string& unused, std::istream& theStream) {
